Updated TimeSeriesGraphClass
Added some logging to TimeSeriesGraphClass (first draft). The increment and decrement functions were a little difficult to read/understand and therefore the logging is slightly barebones in those cases. There might need to be some small changes to this file in the future as well (proper use of getters etc).